refactor/move login logic to hook
