+++ apps/web/src/router.ts
@@ -1,6 +1,6 @@
 import { createRouter, createWebHistory } from 'vue-router';
 import HomeView from '@/pages/HomePage.vue';
-import LoginView from '@/pages/LoginPage.vue';
+import LoginView from '@/pages/LoginPage/LoginPage.vue';
 
 const router = createRouter({
     history: createWebHistory(import.meta.env.BASE_URL),
